requirements.txt

tensorflow==2.15.0
flask==3.0.0
opencv-python==4.8.1.78
pillow==10.1.0
numpy==1.24.3
pandas==2.1.4
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
gunicorn==21.2.0
python-dotenv==1.0.0
werkzeug==3.0.1
flask-cors==4.0.0
pytest==7.4.3
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
